extract types to types file, extract hook to file, move subcomponents into directories
