+++ packages/components/src/Combobox/Combobox.tsx
@@ -1,5 +1,4 @@
-import React, { ReactElement, ReactNode } from "react";
-import { useAssert } from "@jobber/hooks/useAssert";
+import React from "react";
 import { ComboboxContent } from "./components/ComboboxContent";
 import { ComboboxAction } from "./components/ComboboxAction";
 import { ComboboxContextProvider } from "./ComboboxProvider";
@@ -7,27 +6,14 @@ import {
   ComboboxTriggerButton,
   ComboboxTriggerChip,
 } from "./components/ComboboxTrigger";
-
-export interface ComboboxProps {
-  readonly children: ReactElement | ReactElement[];
-
-  /**
-   * When `true`, `Combobox` will allow for multiple selections
-   *
-   * @default false
-   */
-  readonly multiSelect?: boolean;
-}
-
-export const COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE =
-  "Combobox can only have one Trigger element";
-export const COMBOBOX_REQUIRED_CHILDREN_ERROR_MESSAGE =
-  "Combobox must have a Trigger and Combobox.Content element";
+import { ComboboxProps } from "./Combobox.types";
+import { useComboboxValidation } from "./hooks/useComboboxValidation";
 
 export const Combobox = (props: ComboboxProps): JSX.Element => {
   const { contentElement, triggerElement } = useComboboxValidation(
     props.children,
   );
+
   return (
     <ComboboxContextProvider multiselect={props.multiSelect}>
       {triggerElement}
@@ -35,51 +21,6 @@ export const Combobox = (props: ComboboxProps): JSX.Element => {
     </ComboboxContextProvider>
   );
 };
-
-function useComboboxValidation(children: ReactNode): {
-  triggerElement: ReactNode;
-  contentElement: ReactNode;
-} {
-  const childrenArray = React.Children.toArray(children);
-  let triggerElement: ReactNode,
-    contentElement: ReactNode,
-    multipleTriggersFound = false;
-
-  childrenArray.forEach(child => {
-    if (isTriggerElement(child)) {
-      if (triggerElement) {
-        multipleTriggersFound = true;
-      }
-      triggerElement = child;
-    }
-    if (isContentElement(child)) {
-      contentElement = child;
-    }
-  });
-
-  useAssert(multipleTriggersFound, COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE);
-
-  useAssert(
-    !triggerElement || !contentElement,
-    COMBOBOX_REQUIRED_CHILDREN_ERROR_MESSAGE,
-  );
-
-  return {
-    contentElement,
-    triggerElement,
-  };
-}
-
-function isTriggerElement(child: ReactNode): boolean {
-  return (
-    React.isValidElement(child) &&
-    (child.type === ComboboxTriggerButton || child.type === ComboboxTriggerChip)
-  );
-}
-
-function isContentElement(child: ReactNode): boolean {
-  return React.isValidElement(child) && child.type === ComboboxContent;
-}
 
 Combobox.TriggerButton = ComboboxTriggerButton;
 Combobox.TriggerChip = ComboboxTriggerChip;
